+++ systems/coc7-rolls.js
@@ -45,7 +45,7 @@ export class CoC7Rolls extends BaseRolls {
     defaultRequest(app) {
         let allPlayers = (app.tokens.filter(t => t.actor?.hasPlayerOwner).length == app.tokens.length);
         //if all the tokens have zero hp, then default to death saving throw
-        let allZeroHP = app.tokens.filter(t => getProperty(t.actor, "system.attributes.hp.value") == 0).length;
+        let allZeroHP = app.tokens.filter(t => foundry.utils.getProperty(t.actor, "system.attributes.hp.value") == 0).length;
         return (allZeroHP == app.tokens.length && allZeroHP != 0 ? 'misc:death' : null) || (allPlayers ? 'skill:prc' : null);
     }*/