+++ systems/ose-rolls.js
@@ -11,7 +11,7 @@ export class OSERolls extends BaseRolls {
         ].concat(this._requestoptions);
 
         /*
-        this._defaultSetting = mergeObject(this._defaultSetting, {
+        this._defaultSetting = foundry.utils.mergeObject(this._defaultSetting, {
             stat1: "ac.value"
         });*/
     }