+++ systems/pf2e-rolls.js
@@ -15,7 +15,7 @@ export class PF2eRolls extends BaseRolls {
         ].concat(this._requestoptions);
 
         /*
-        this._defaultSetting = mergeObject(this._defaultSetting, {
+        this._defaultSetting = foundry.utils.mergeObject(this._defaultSetting, {
             stat2: "attributes.perception.value + 10"
         });*/
     }
@@ -239,8 +239,8 @@ export class PF2eRolls extends BaseRolls {
     }
 
     async checkXP(actor) {
-        if (setting("send-levelup-whisper") && actor.system.details.xp.value >= actor.system.details.xp.max) {
-            const level = parseInt(getProperty(actor, "system.details.level.value")) + 1;
+        if (setting("send-levelup-whisper") && game.user.isTheGM && actor.system.details.xp.value >= actor.system.details.xp.max) {
+            const level = parseInt(foundry.utils.getProperty(actor, "system.details.level.value")) + 1;
             const html = await renderTemplate("./modules/monks-tokenbar/templates/levelup.html", { level: level, name: actor.name, xp: actor.system.details.xp.value });
             ChatMessage.create({
                 user: game.user.id,
@@ -255,7 +255,7 @@ export class PF2eRolls extends BaseRolls {
 
     getValue(actor, type, key) {
         let prop = type == "skill" ? "skills" : type == "save" ? "saves" : "attributes";
-        let value = getProperty(actor, prop + "." + key + ".mod");
+        let value = foundry.utils.getProperty(actor, prop + "." + key + ".mod");
         return value;
     }
 }